료칸 등록 컴포넌트 모듈화, 버튼 상태관리 리덕스 적용
