+++ next_reservation/components/Register/RegisterFooter/RegisterFooter.tsx
@@ -1,18 +1,16 @@
 import Link from 'next/link';
-import React, { useCallback } from 'react';
+import React, { useEffect, useState } from 'react';
+import { useSelector } from '../../../store';
 import RegisterFooterStyle from '../../../styles/components/Register/RegisterFooter';
 
 interface IProps {
-  isValild: boolean;
   previousHref: string;
   nextHref: string;
 }
 
-const RegisterFooter: React.FC<IProps> = ({
-  isValild,
-  nextHref,
-  previousHref,
-}) => {
+const RegisterFooter: React.FC<IProps> = ({ nextHref, previousHref }) => {
+  const isValid = useSelector((selector) => selector.registerIsValid.isValid);
+
   return (
     <RegisterFooterStyle className="h-20 w-1/2 p-10 absolute left-1/2 bottom-10 flex items-center justify-between">
       <Link href={previousHref || ''} passHref>
@@ -21,7 +19,7 @@ const RegisterFooter: React.FC<IProps> = ({
         </button>
       </Link>
       <Link href={nextHref || ''} passHref>
-        <button className="btn btn-primary" disabled={isValild}>
+        <button className="btn btn-primary" disabled={!isValid}>
           <a>다음</a>
         </button>
       </Link>
