+++ next_reservation/pages/room/register/ryokan.tsx
@@ -1,11 +1,18 @@
 import React from 'react';
 import { GetServerSideProps, NextPage } from 'next';
 import RegisterRyokan from '@/components/Register/RegisterRyokan/RegisterRyokan';
-
-type serverSidePropsType = { priviousHref: string };
+import RegisterRyokanType from '@/components/Register/RegisterRyokanType/RegisterRyokanType';
+import { serverSidePropsType } from 'types/registRyokanServerSidePropType';
 
 const ryokan: NextPage<serverSidePropsType> = (props: serverSidePropsType) => {
-  return <RegisterRyokan priviousHref={props.priviousHref} />;
+  return (
+    <RegisterRyokan
+      priviousHref={props.priviousHref}
+      nextHref="/room/register/bedrooms"
+    >
+      <RegisterRyokanType />
+    </RegisterRyokan>
+  );
 };
 
 export const getServerSideProps: GetServerSideProps = async (context) => {
