+++ next_reservation/store/index.ts
@@ -8,12 +8,14 @@ import {
 import { configureStore, combineReducers } from '@reduxjs/toolkit';
 import userSignInAndUp from './userSignInAndUp';
 import registerRyokan from './registerRyokan';
+import registerFormIsValid from './registerFormIsValid';
 import rootSaga from './sagas';
 
 // 여러개의 리듀서 컴바인
 const rootReducer = combineReducers({
   user: userSignInAndUp.reducer,
   registerRyokan: registerRyokan.reducer,
+  registerIsValid: registerFormIsValid.reducer,
 });
 
 // 스토어 타입
